+++ model/group.go
@@ -5,7 +5,7 @@ import "time"
 type Group struct {
 	ID          string
 	Name        string
-	AccessLevel int
+	AccessLevel float32
 	CreateDate  time.Time
 	UpdateDate  time.Time
 }
